fix/unregister (#52)
* feat: add logs for scrim events

* refactor: unregister command and kick-team interaction

* style: remove line

* feat: participant role, slot logs

add participant role assignment on slot assignment and deletion
add logs for slot events

* fix: use assignslot logic

* fix: reuse unassign slot logic

* fix: order of logging event

* fix: team kick action

* fix: ban team logs

* feat: log ban event

* fix: order of ban event

* feat: add log for registration closed

* chore: disable unregister command

* feat: delete roles on delete scrim

---------

+++ src/commands/esports/delete.ts
@@ -42,6 +42,7 @@ export default class ScrimDelete extends Command {
   async execute(interaction: ChatInputCommandInteraction<"cached">) {
     const scrimId = interaction.options.getInteger("id");
     let scrim: Scrim | null = null;
+
     if (scrimId === null) {
       scrim = await prisma.scrim.findFirst({
         where: { adminChannelId: interaction.channelId },
@@ -78,13 +79,19 @@ export default class ScrimDelete extends Command {
       scrim.participantsChannelId,
       scrim.discordCategoryId,
     ];
+
     await Promise.allSettled(
       deletableChannels.map((channelId) =>
         rest.delete(Routes.channel(channelId)),
       ),
     );
-
+    const guild = await this.client.guilds.fetch(scrim.guildId);
+    if (!guild) {
+      return;
+    }
+    suppress(this.client.rolemanageService.deleteParticipantRole(guild, scrim));
     await prisma.scrim.delete({ where: { id: scrim.id } });
+
     await suppress(
       interaction.editReply({
         content: `Scrim with ID ${scrim.id} has been deleted.`,
